Fix incorrect clock base for SCTE-35 duration field
The generation of the SCTE-35 packet was specifying the duration
field in 0.1 increments (which is what it got from the SCTE-104
parser).  However the SCTE-35 duration is a 33-bit value and must
be specified based on a 90 KHz clock.

See SCTE-35-2016 Section 9.4.2: break_duration().

Adjust the clock base to be correct per the SCTE-35 spec.