Fix case where installed copies of headers/libs took precendence over local copies
Because of the order for the CFLAGS and LDFLAGS processing, copies
of the headers and libraries which might be installed on the system
would take precendence over versions in the actual source tree.  This
would cause all sorts of misleading problems during development
where the wrong headers and libs were being built into the test apps.

Update the automake directives to properly use the versions in
the actual source tree rather than copies which might be installed.

This also fixes a case where changes to src/libklvanc.a weren't
resulting in the binaries in tools/ being rebuilt.