Support other splice types in 35->104, and move serialization to klvanc
The previous code only supported the splice_insert Operation.  Lay the
groundwork to support other SCTE-104 message types based on the
input SCTE-35 segment.

Also, move the code which serializes an MOM into a binary blob into
libklvanc, since this is all specific to SCTE-104 and has no
bearing on SCTE-35.  This also allows us to reuse this functionality
within libklvanc itself for test cases where we want to deserialize
a 104 packet, re-serialize it, and then compare the result to
validate the encoder and decoder.