Break out to/from 104 support into separate files.
The scte35.c was getting too unwieldy, so create separate files for
the core SCTE-35 functionality, and then separate files for
conversion to/from SCTE-104.

This will also make it easier at some point to build without
libklvanc for projects that only are doing SCTE_35 and are not
concerned with converting back/forth with SCTE-104.