Rework generation of SCTE-35 descriptors
Before we were just blindly copying the array of bytes from input
to output, which prevents us from working with cases where the
descriptor needs to be generated (e.g. when bridging from SCTE-104
to SCTE-35).  Refactor the code such that the SCTE-35 decoder
creates an intermediate struct to represent the descriptor in the
splice_info struct, and separately have the encoder use that struct
when serializing the final SCTE-35 section.

Note this commit only implements the DTMF descriptor, but lays
the groundwork for all the other descriptors.