Fix API breakage from when we stopped generating SMPTE 2010 packets
When we refactored the API last year such that the SCTE-104 payload
generated was no longer wrapped in a SMPTE 2010 VANC packet, we
needed API callers to start explicitly generating the 2010 packet
if desired.  Add the additional call to create the 2010 packet.

Because klvanc_convert_SCTE_104_packetbytes_to_SMPTE_2010() expects
a klvanc context, we had to move that to the top of the function,
and because it always sets the callback, that needed to no longer
be in an #ifdef DEBUG_RENDER_104.  There is no functional change
though - the user still needs to #define DEBUG_RENDER_104 if they
want to see the SCTE-104 decoded output.