Fix rounding error in duration conversion for SCTE35 to 104.
Because the SCTE-35 duration is expressed on a 90KHz clock but the
SCTE-104 duration is in 0.1 second increments, the conversion
arithmetic can result the 104 duration always being rounded down.

For example, I saw a case where the math ended up with a value
of 27.994 seconds, which would always be rounded down to 27.9 seconds
instead of 28.0 (i.e. off by nearly 100ms).

Tweak the math to round up when appropriate.