Take into account pts_adjustment field when converting SCTE 35 to 104
The existing conversion routine didn't use the value in the splice_info
pts_adjustment field.  As a result, if the upstream source had reclocked
the PTS timing during a transcode (or if the encoder just happens to set
it for whatever reason), we weren't compensating for the adjusted PTS
in determining the preroll for any resulting SCTE-104 splice_request
or time_signal packets.

Subtract out any PTS adjustment prior to using the PTS value for
computing pre-roll values.