Don't treat bandwidth_reservation as an error when converting to 104
The SCTE-104 protocol doesn't have a message that corresponds to
the SCTE-35 bandwidth_reservation message, and the SCTE-35 spec
indicates such messages should just be dropped.

Because bandwidth_reservation messages will be sent continuously
by some encoders, it shouldn't be treated as an error condition,
but rather as a success that results in no SCTE-104 payload to
inject.