Ignore descriptors with identifer others than "CUEI"
SCTE-35:2016 Sec 10.1 states "Any receiving equipment should skip
any descriptors with unknown identifiers or unknown descriptor tags".

We actually found such a case where the descriptor tag was 0x02
(segmentation_descriptor), but the identifer was not CUEI and the
structure of the private bytes didn't match up to the spec.

Ignore such descriptors entirely for now.  In the longer term we
can look at moving the identifier check up into the caller, and
then support the ability to pass these as private descriptors.