Don't bail out if reserved bits are zero
While the spec says that reserved bits should be set to zero, it
also says that receivers should explicitly ignore the values
(presumably because future versions might use those bits).

Hence, don't bail out if a packet is presented where reserved
bits are zero.  In fact, the comcast_gots_test1 and
comcast_gots_test1 samples in tools/parse.c exhibit this behavior.

The language in Sec SCTE-35:2019 5.2 is specifically:

"Unless otherwise specified, all reserved bits shall be set to '1'
and this field shall be ignored by receiving equipment."

We might at some point want to add some specific developer mode
which cross-checks to ensure reserved bits are set to '1' in payloads
we generate, but our general parser shouldn't be bailing out if the
we receive a packet where reserved bits are zero.