Add support for dumping contents of descriptors
In order to aid in debugging, implement the routines to dump out
the contents of descriptors tied to SCTE-35 messages.

Note that the lookup tables for various values are currently
defined as static as I haven't decided whether they should be part
of the public API (or if the names should be changed to be more
consistent).

Verified by manually comparing the output for the test vectors as
documented in SCTE-35:2019 Section 14.