Extend test tool to compare original blobs to output
In order to exercise the parser and generator, after we've parsed
the packet feed it back into the generator and compare the resulting
binary blob against the original.

Note that with this commit, the results will show a number of
failures (i.e. not all tests pass currently).  The pass rate will
increase with subsequent patches to the parser/generators in the
library itself.