Add support for program_segmentation_flag (i.e. component mode)
Implement support for the less common case where a segmentation
descriptor has program_segmentation_flag set to zero.  We have a
test case in the suite which meets this criteria and this patch
results in that test case now passing.

Note that this support is for the SCTE-35 parts, but we still don't
support 35->104 and vice versa for component mode.