Refactor the way supplemental descriptors are serialized/deserialized
Address a number of inconsistencies found in handling descriptors,
in particular when the identifier isn't CUEI (i.e. non-SCTE descriptors).

Because the identifier, length, and descriptor type are mandatory
for all messages, move their processing further up rather than handing
off the data to the specific message types.  Thus we're now setting
the identifier and descriptor length in the main routine and only
handing off the private bytes to the specific message functions.

Also, rework the decision making such that we only call the specific
message functions if, and only if, the identifier is CUEI.  Otherwise
use the generic routine for adding opaque/undocumented descriptors.
This avoids cases where the identifier is not CUEI but the descriptor
type happens to overlap with one of the documented CUEI types.