Fix 104->35 preroll rollover & criteria.
While splice_insert messages with zero preroll are supposed to be
given a pts of zero, and the type set to immediate, time_signals
do not have an equivalent concept, thus the pts should be provided
regardless of the preroll specified in SCTE 104.

A guard against exceeding the maximum PTS value of 33 bits was also
included for time_signal() and splice_insert() translations from
SCTE 104.