Fix segfault in scte104 testcase when passed bandwidth reservation
The SCTE-35 bandwidth reservation message doesn't have an SCTE-104
equivalent, so the conversion routine returns success but doesn't
provide back a payload.  This can cause the caller to crash if it
tries to free() the output buffer.

Check the output buffer before trying to free it.

This bug wasn't caught until refactoring the test cases such that
all test cases are run through all the different tools.