Make inclusion of libklvanc optional
There are use cases where people might want to use klscte35 but
don't need anything involving SCTE-104, and thus don't want/need
to build libklvanc.

Make the inclusion of libklvanc optional (defaults to include it
if available).

Note that because the public libklscte35 header no longer includes
<libklvanc/vanc.h>, applications that want to use it need to have
the include themselves.  This could be considered a breaking change
in the API, but I suspect most applications already explicitly
include it.